requirements.txt
PyMuPDF==1.26.1
pandas==2.3.0
requests==2.32.3
ffmpeg-python==0.2.0
python-docx==1.1.2
Django==3.2.3
pillow==11.2.1
moviepy==2.2.1
playwright==1.52.0
openai==1.84.0
lxml==5.4.0
google==3.0.0
protobuf==6.31.1
PyMySQL==1.1.1
pathlib==1.0.1